+++ internal/api/routes.go
@@ -7,8 +7,8 @@ import (
 	"github.com/edulinq/autograder/internal/api/core"
 	"github.com/edulinq/autograder/internal/api/courses"
 	"github.com/edulinq/autograder/internal/api/lms"
-	"github.com/edulinq/autograder/internal/api/server"
 	"github.com/edulinq/autograder/internal/api/submissions"
+	"github.com/edulinq/autograder/internal/api/users"
 )
 
 var baseRoutes = []*core.Route{
@@ -27,8 +27,8 @@ func GetRoutes() *[]*core.Route {
 	routes = append(routes, *(admin.GetRoutes())...)
 	routes = append(routes, *(courses.GetRoutes())...)
 	routes = append(routes, *(lms.GetRoutes())...)
-	routes = append(routes, *(server.GetRoutes())...)
 	routes = append(routes, *(submissions.GetRoutes())...)
+	routes = append(routes, *(users.GetRoutes())...)
 
 	return &routes
 }
