--- internal/api/server/users/main_test.go
@@ -1,12 +1,0 @@
-package users
-
-import (
-	"testing"
-
-	"github.com/edulinq/autograder/internal/api/core"
-)
-
-// Use the common main for all tests in this package.
-func TestMain(suite *testing.M) {
-	core.APITestingMain(suite, GetRoutes())
-}
