+++ internal/api/users/tokens_delete_test.go
@@ -8,7 +8,7 @@ import (
 	"github.com/edulinq/autograder/internal/util"
 )
 
-func TestTokensCreate(test *testing.T) {
+func TestTokensDelete(test *testing.T) {
 	db.ResetForTesting()
 	defer db.ResetForTesting()
 
@@ -18,8 +18,15 @@ func TestTokensCreate(test *testing.T) {
 	}
 
 	initialTokenCount := len(user.Tokens)
+	if initialTokenCount == 0 {
+		test.Fatalf("Test user has no tokens.")
+	}
 
-	response := core.SendTestAPIRequest(test, core.NewEndpoint("server/users/tokens/create"), nil)
+	args := map[string]any{
+		"token-id": user.Tokens[0].ID,
+	}
+
+	response := core.SendTestAPIRequest(test, core.NewEndpoint("users/tokens/delete"), args)
 	if !response.Success {
 		test.Fatalf("Response not successful: '%s'.", util.MustToJSONIndent(response))
 	}
@@ -31,7 +38,7 @@ func TestTokensCreate(test *testing.T) {
 
 	newTokenCount := len(user.Tokens)
 
-	if newTokenCount != (initialTokenCount + 1) {
-		test.Fatalf("Incorrect token count. Expected: %d, Found: %d.", (initialTokenCount + 1), newTokenCount)
+	if newTokenCount != (initialTokenCount - 1) {
+		test.Fatalf("Incorrect token count. Expected: %d, Found: %d.", (initialTokenCount - 1), newTokenCount)
 	}
 }
