+++ testdata/course-languages/course.json
@@ -1,6 +1,6 @@
 {
     "id": "course-languages",
-    "name": "Course Using Different Languages",
+    "name": "Course Using Different Languages.",
     "source": {
         "type": "nil"
     },
